Update headers and fix border
